[Fortran/gfortran] Better handling of unsupported gfortran options
A number of tests use gfortran-specific options which are either not supported
in flang, or for which a suitable equivalent has not yet been enabled. These
options used to be filtered out in CMakeLists.txt when configuring the test
suite. This has been moved to the test configuration script to allow for more
flexibility (since cmake is a terrible language in which to do anything even
remotely sophisticated).

Doing this has allowed us to perform wildcard matches for options that expose
gfortran's internal structure which is obviously not relevant for flang. This
means that in the future, we do not have to constantly update the filter with
these options. We can now also distinguish between the option name and its
value, so we can filter simply based on the option spelling.

The test configuration files have been regenerated after this change. Some
tests fail with this updated configuration and they have been disabled.