[gfortran] Fixed cray_pointers_2.f90 aliasing issues. (#252)
The test started failing after https://github.com/llvm/llvm-project/pull/140803
The issue is that the Cray pointee and the underlying storage
are both accessed in subroutines violating Fortran aliasing rules
for the dummy arguments.

GCC tried to "solve" that by using `-fno-inline` option, but
it should be better to make the aliasing explicit with proper `TARGET`
attribute placement.

* Updated comments.