[flang][tests] Disable Fortran/gfortran/regression/finalize_38.f90
One test case in finalize_38.f90 expects that an allocatable
component of an expression result will not be finalized.  The
flang-new compiler is being updated by the pull request

  https://github.com/llvm/llvm-project/pull/153509

to ensure finalization of allocatables when they are deallocated,
since this is an event that triggers finalization (F'2023 7.5.6.3p2)
on a finalizable (3.71) data entity (3.41).