[HIP] Add a __builtin_memset test (#276)
This is mostly analogous to the existing memmove test: __builtin_memset is
called on the device with different lengths, in different address spaces, at
different dynamic alignments, and with constant and dynamic operands. The
result is compared to a reference computation on the host to ensure functional
correctness of the memset lowering.

Also adds a missing standard lib include to the memmove test.

For SWDEV-543208.