[HIP] Increase blender perf diff threshold
This parameter sets how much difference in runtime the blender workloads
are allowed to show when run in the testing.
In reality it was advised to set this to 15% instead of the 5% shown
here.

With this patch, we increase the default threshold to 10% to see if this
gives us enough wiggle room out of the box to not see any runtime
related flakiness.