Fix remaining benchmarks with GCC 15
In GCC 15 the default C standard was changed to gnu23, so similar to #285 this explicitly sets -std=gnu17 on benchmarks that are known to fail to build with GCC 15 because of this.

Whilst we're here, this also goes through and replaces -std=X flags with the more CMake-native C_STANDARD property.